Modifications

+++ app/views/back/admin.php
@@ -1,28 +1,37 @@
 <?php ob_start(); ?>
 
-<?php $title = 'Gestion des articles'; ?>
-<?php $titleCat = 'Gestion des articles'; ?>
+<?php
+session_start();
+if (isset($_SESSION['pseudo'])) {?>
 
-<a href="index.php?p=newPost"><input name="ButtonAddPost" value="Nouvel article" type="button" /></a>
+	<p>hello <?= $_SESSION['pseudo'] ?></p>
+	<?php $title = 'Gestion des articles'; ?>
+	<?php $titleCat = 'Gestion des articles'; ?>
 
-	<?php foreach ($posts as $post): ?>
+	<a href="index.php?p=newPost"><input name="ButtonAddPost" value="Nouvel article" type="button" /></a>
 
-		<div>
-			<h3> <?= htmlspecialchars($post->getTitle()); ?> </h3>
-			<p>
-				<em>le <?= $post->getDate();?></em> par <?= $post->getAuthor(); ?>
-			</p>
-				
-			<p>
-				<?= substr($post->getPost(), 0, 100); ?>
-			</p>
-		</div>
-		<div>
-			<a href="index.php?p=deletePost&amp;id=<?= $post->getId() ?>"><input name="ButtonDeletePost" id="deletePostJS" value="Supprimer l'article" type="submit" /></a>
-			<a href="index.php?p=editPost&amp;id=<?= $post->getId() ?>"><input name="ButtonEditPost" value="Modifier l'article" type="submit" /></a>
-		</div>
+		<?php foreach ($posts as $post): ?>
 
-	<?php endforeach; ?>
+			<div>
+				<h3> <?= htmlspecialchars($post->getTitle()); ?> </h3>
+				<p>
+					<em>le <?= $post->getDate();?></em> par <?= $post->getAuthor(); ?>
+				</p>
+					
+				<p>
+					<?= strip_tags(substr(($post->getPost()), 0, 200)); ?>
+				</p>
+			</div>
+			<div>
+				<a href="index.php?p=deletePost&amp;id=<?= $post->getId() ?>"><input name="ButtonDeletePost" id="deletePostJS" value="Supprimer l'article" type="submit" /></a>
+				<a href="index.php?p=editPost&amp;id=<?= $post->getId() ?>"><input name="ButtonEditPost" value="Modifier l'article" type="submit" /></a>
+			</div>
+
+		<?php endforeach;
+
+} else {
+	header('Location: index.php?p=connexion');
+} ?>
 
 <?php $content = ob_get_clean(); ?>
